Rename `package db` to `package datasource`
