+++ cmd/out.go
@@ -24,7 +24,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/k1LoW/tbls/db"
+	"github.com/k1LoW/tbls/datasource"
 	"github.com/k1LoW/tbls/output"
 	"github.com/k1LoW/tbls/output/dot"
 	"github.com/k1LoW/tbls/output/json"
@@ -50,7 +50,7 @@ var outCmd = &cobra.Command{
 	},
 	Run: func(cmd *cobra.Command, args []string) {
 		dsn := args[0]
-		s, err := db.Analyze(dsn)
+		s, err := datasource.Analyze(dsn)
 		if err != nil {
 			printError(err)
 			os.Exit(1)
