+++ datasource/datasource_test.go
@@ -1,4 +1,4 @@
-package db
+package datasource
 
 import (
 	"testing"
